\name{hdbma-package}
\alias{hdbma-package}
\docType{package}
\title{
\packageTitle{hdbma}
}
\description{
\packageDescription{hdbma}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{hdbma}
\packageIndices{hdbma}
The main function is hdbma to perform the Bayesian mediation anlysis with adaptive Laplace priors.
}
\author{
\packageAuthor{hdbma}

Maintainer: \packageMaintainer{hdbma}
}
\references{
Yu, Q., Hagan, J., Wu, X., Richmond-Bryant, J., and Li, B., 2023, High-Dimensional Bayesian Mediation Analysis with Adaptive Laplace Priors. Submitted.
}

\keyword{high-dimensional Bayesian mediation analysis}
\examples{
#See examples at summary.hdbma.
}
